+++ ycc_hull/api/members.py
@@ -1,11 +1,12 @@
 """
 Member API endpoints.
 """
+from collections.abc import Sequence
 from datetime import date
-from typing import Optional, Sequence
+from typing import Optional
 
 from fastapi import APIRouter, Depends
-from ycc_hull.auth import AuthInfo, auth
+from ycc_hull.auth import User, auth
 
 from ycc_hull.controllers.members_controller import MembersController
 from ycc_hull.api.errors import (
@@ -18,11 +19,11 @@ api_members = APIRouter(dependencies=[Depends(auth)])
 
 @api_members.get("/api/v0/members")
 async def members_get(
-    year: Optional[int] = None, auth_info: AuthInfo = Depends(auth)
+    year: Optional[int] = None, user: User = Depends(auth)
 ) -> Sequence[MemberPublicInfoDto]:
     current_year = date.today().year
 
-    if year != current_year and not (auth_info.admin or auth_info.committee_member):
+    if year != current_year and not (user.admin or user.committee_member):
         raise create_http_exception_403(
             f"You do not have permission to list members for {year}"
         )
@@ -36,8 +37,8 @@ async def membership_types_get() -> Sequence[MembershipTypeDto]:
 
 
 @api_members.get("/api/v0/users")
-async def users_get(auth_info: AuthInfo = Depends(auth)) -> Sequence[UserDto]:
-    if not auth_info.admin:
+async def users_get(user: User = Depends(auth)) -> Sequence[UserDto]:
+    if not user.admin:
         raise create_http_exception_403("You do not have permission to list users")
 
     return await MembersController.find_all_users()